Split `data.scala` into many smaller files
