+++ src/core/project.scala
@@ -1,0 +1,39 @@
+package fury.core
+
+import fury.model._, fury.strings._
+
+import mercator._
+
+import scala.util._
+import scala.collection.immutable._
+
+object Project {
+  implicit val msgShow: MsgShow[Project]       = v => UserMsg(_.project(v.id.key))
+  implicit val stringShow: StringShow[Project] = _.id.key
+  implicit def diff: Diff[Project]             = Diff.gen[Project]
+
+  def available(projectId: ProjectId, layer: Layer): Boolean =
+    !layer.projects.toOption.to[List].flatten.findBy(projectId).isSuccess
+}
+
+case class Project(id: ProjectId,
+                   modules: SortedSet[Module] = TreeSet(),
+                   main: Option[ModuleId] = None,
+                   license: LicenseId = License.unknown,
+                   description: String = "",
+                   compiler: Option[ModuleRef] = None) {
+
+  def apply(module: ModuleId): Try[Module] = modules.findBy(module)
+  def moduleRefs: List[ModuleRef] = modules.to[List].map(_.ref(this))
+  def mainModule: Try[Option[Module]] = main.map(modules.findBy(_)).to[List].sequence.map(_.headOption)
+
+  def compilerRefs: List[ModuleRef] =
+    modules.to[List].collect { case m if m.kind == Compiler => m.ref(this) }
+
+  def unused(moduleId: ModuleId) = modules.findBy(moduleId) match {
+    case Success(_) => Failure(ModuleAlreadyExists(moduleId))
+    case _          => Success(moduleId)
+  }
+
+  def allRepoIds: Set[RepoId] = modules.flatMap(_.sources).collect { case ExternalSource(repoId, _) => repoId }
+}
